backend/requirements.txt
crewai
crewai[tools]
langchain-google-genai
google-search-results
firebase-admin
python-dotenv
